feat(vote): get votes in correct order based on the order created at the poll
